ai-management/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
aiohttp==3.9.1
redis==5.0.1
aioredis==2.0.1
